requirements.txt
Django==2.2.1
django-filter==2.1.0
django-macros-url==0.4.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
Markdown==3.1.1
psycopg2-binary==2.8.2
python-memcached==1.59
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
